builtin/pack-objects.c: freshen objects from existing cruft packs
Since 61568efa95 (builtin/pack-objects.c: support `--max-pack-size` with
`--cruft`, 2023-08-28), 'pack-objects' supports writing multiple cruft
packs when invoked with `--cruft` and the `--max-pack-size` flag.

This process is typically driven by 'repack', where 'repack' will
combine cruft packs smaller than the given threshold. However, this
process breaks down when we attempt to freshen an object packed in an
earlier cruft pack that is larger than the threshold and thus will
survive the repack.

When this is the case, any objects in cruft pack(s) larger than the
threshold are unable to be freshened. This is because we avoid writing
them in the new cruft pack entirely, for a couple of reasons.

 1. When enumerating packed objects via 'add_objects_in_unpacked_packs()'
    we pass the SKIP_IN_CORE_KEPT_PACKS, which is used to avoid looping
    over the packs we're going to retain (which are marked as kept
    in-core by 'read_cruft_objects()').

    This means that we will avoid enumerating additional packed copies
    of objects found in any cruft packs which are larger than the given
    size threshold. Thus there is no opportunity to call
    'create_object_entry()' whatsoever.

 2. We likewise will discard the loose copy (if one exists) of any
    unreachable object packed in a cruft pack that is larger than the
    threshold. Here our call path is 'add_unreachable_loose_objects()',
    which uses the 'add_loose_object()' callback.

    That function will eventually land us in 'want_object_in_pack()'
    (via 'add_cruft_object_entry()'), and we'll discard the object as it
    appears in one of the packs which we marked as kept in-core.

This means in effect that it is impossible to freshen an unreachable
object after it has been packed in a cruft pack larger than the given
threshold.

Instead, we should pack additional copies of any unreachable object that
we want to freshen only when we can find a copy of it with an mtime more
recent than the most recent mtime for that object among any copies of it
in cruft packs.

This is sub-optimal in the sense that it requires keeping an additional
copy of unreachable objects upon freshening, but we don't have a better
alternative without the ability to make in-place modifications to
existing *.mtimes files.

In order to implement this, we have to adjust the behavior of
'want_found_object()'. When 'pack-objects' is told that we're *not*
going to retain any cruft packs (i.e. the set of packs marked as kept
in-core does not contain a cruft pack), the behavior is unchagned.

But when there *is* at least one cruft pack that we're holding onto, it
is no longer sufficient to reject a copy of an object found in that
cruft pack for that reason alone. In this case, we only want to reject a
candidate object when copies of that object can be found in a cruft pack
if the most recent mtime of any existing cruft copy of that object is
more recent than the one we're debating whether to pack.

To do this, keep track of whether or not we have any cruft packs in our
in-core kept list with a new 'ignore_packed_keep_in_core_has_cruft'
flag. When we end up in this new special case, we replace a call to
'has_object_kept_pack()' to 'find_recent_kept_pack_entry()', and only
reject objects when we have a copy in an existing cruft pack with a more
recent mtime (in which case "freshening" would be redundant).

Signed-off-by: Taylor Blau <me@ttaylorr.com>