format-patch: split Message-ID generation in two
gen_message_id() has two callers: one which generates the cover letter's
Message-ID, and another which generates Message-IDs for the individual
patches themselves.

So the second argument to gen_message_id() is either "cover", or the
result of calling oid_to_hex(). The latter case makes calling
gen_message_id() slightly awkward. It would be more convenient if the
caller could simply pass a pointer to the 'struct object_id' directly
instead of first having to call oid_to_hex().

Split gen_message_id() into two functions: gen_message_id_literal() and
the original gen_message_id(). The former is used when we have a literal
string that we want to use as part of the Message-ID (like "cover"). The
latter takes a pointer to a 'struct object_id', and implements itself by
calling the former.

This prepares us nicely for the following commit, which will make the
latter case slightly more complex.

Signed-off-by: Taylor Blau <me@ttaylorr.com>