format-patch: use abbreviated OIDs for threaded Message-IDs
When formatting a patch with threading, we use the object's full
hex-encoded OID as part of the Message-ID, like so:

    $ git.compile format-patch HEAD^ --stdout --thread | grep '^Message-ID:'
    Message-ID: <fa0e9e06067931c2fdfe6c341c4f66e66e590c3c.1742341797.git.me@ttaylorr.com>

, which results in quite long Message-IDs.

While there's nothing inherently wrong with having a long Message-ID
(RFC 5322 §2.1.1 permits header lines to be as much as 998 characters
long), it can make linking to individual messages on mailing list
archives awkward, particularly when the reference itself is
line-wrapped.

There's no practical reason why we can't abbreviate the OID when
generating a Message-ID as above. The likelihood of a Message-ID
collision is practically unchanged, since we still include the current
timestamp as part of the Message-ID.

Now, the Message-ID header is much more manageable, and looks something
like:

    $ git.compile format-patch HEAD^ --stdout --thread | grep '^Message-ID:'
    Message-ID: <fa0e9e0606.1742341797.git.me@ttaylorr.com>

which is much less awkward to write.

Signed-off-by: Taylor Blau <me@ttaylorr.com>