midx-write.c: compute MIDX pseudo-pack order directly
The function midx_pack_order() computes the pseudo-pack order for
objects in a multi-pack index when writing a multi-pack bitmap. It has a
single caller (in 'write_midx_internal()') which does the following:

    ctx.pack_order = midx_pack_order(&ctx);

It is somewhat awkward to have midx_pack_order() return a freshly
allocated array, only to assign it to a field which midx_pack_order()
itself can access through the pointer it gets to 'ctx'.

Remove that awkwardness and make the lifetime of the 'pack_order' field
clearer by teaching midx_pack_order() to write its result directly to
the given 'ctx'.

Since this function now directly computes and stores its result, rename
it to 'compute_midx_pack_order()' to more clearly reflect its operation.

Signed-off-by: Taylor Blau <me@ttaylorr.com>