blame-tree: drop 'blametree.revopts' configuration
We got this way back in d68cdc3b61 (blame-tree: add config to add to
revopts, 2015-09-14), which explains that it is useful to stick
repo-specific traversal limiting in those repository's 'config' files.

But, between commit-graphs, changed-path Bloom filters, and our new
blame-tree traversal, we like don't need this any more. Let's remove it
in the name of simplicity, and keep an eye on the graphs as we roll it
out to make sure that we aren't subtly relying on it (in which case we
could just as easily drop this patch).