blame-tree.c: don't call 'prepare_revision_walk()'
While implementing 0603f6d9c3 (blame-tree: introduce '--go-faster' mode,
2020-07-16), we inserted a call to 'prepare_revision_walk()' before we
started our traversal. This was done when we leveraged the revision
machinery more (at the time, we were leaning on 'log_tree_commit()'
which only worked after calling 'prepare_revision_walk()').

But, we have since dropped 'log_tree_commit()'. However, I didn't look
further into whether or not the 'prepare_revision_walk()' call could be
dropped, too.

It turns out that not only _can_ this call be dropped, it _should_ be
dropped because it involves a complete traversal of the history when we
are bounded by a bottom ^commit (and hence we are limited). In this
case, 'prepare_revision_walk()' calls 'revision.c:limit_list()', which
does a complete traversal.

It turns out that this looks something like:

    $ git blame-tree --go-faster -z --max-depth=1 TIP ^BOUNDARY -- app

where TIP was the tip of 'master' on github/github, and BOUNDARY was the
most recent thing in the blame-tree-cache.

Here's where things get interesting. BOUNDARY happens to be a commit
from a year or so ago. There's a lot of history between now and then.

Running things in a debugger, this ended up getting resolved pretty
quickly. Stepping through the new implementation, we would hang on
'prepare_revision_walk()' for a number of seconds. This all came down to
'limit_list()', and our journey was over. I removed the call, validated
that things still work, and all was fast again.