pack-bitmap: record whether the result was filtered
In a following commit, we will prepare to mark objects for reuse which
are stored as deltas, but whose base object wasn't included in the
output pack (e.g., because the client already has it).

We can't, however, take advantage of that optimization when the
traversal removed objects from the result due to either (a) object
filtering, or (b) --unpacked.

That's because in either case, we can't rely on the principle that "if
an object appears in the 'haves' bitmap, that the client already has
that object", since they may not have it as a result of filtering.

Keep track of whether or not we performed any object filtering during
our traversal in preparation to actually implement thin delta
conversion.

Signed-off-by: Taylor Blau <me@ttaylorr.com>